requirements.txt
langchain
langchain-aws
langchain-community
faiss-cpu
fastapi
uvicorn
python-dotenv
pydantic
PyPDF2
streamlit
requests
langchain_pinecone
rapidfuzz
pandas
boto3
python-dotenv